Simplified time strech augmentation
