+++ train.py
@@ -183,11 +183,9 @@ def main():
     train_batcher = batcher(sampler(train_data))
     val_batcher = batcher(sampler(val_data))
 
-    """
     # Checks if training data sounds right.
-    for i, (train_seq, *_) in enumerate(train_batcher()):
-        save_midi('train_seq_{}'.format(i), train_seq[0].cpu().numpy())
-    """
+    for i, seq in enumerate(train_batcher()[0]):
+        save_midi('train_seq_{}'.format(i), seq.cpu().numpy())
 
     print('Training Sequences:', len(train_data[0]), 'Validation Sequences:', len(val_data[0]))
     print()
